+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * [EACQAPW-5209] MoneyFormatter Tests
 * [EACQAPW-5238] PaymentController Tests
 * [EACQAPW-5262] CardsController Tests
+* [EACQAPW-5218] AddCardController final tests
 
 ### Fixed